+++ src/donut/browser-types.ts
@@ -1,0 +1,1 @@
+export const DONUT_SUPPORTED_BROWSER_TYPES = new Set(['camoufox', 'weyfern', 'wayfern']);